Titles & more grade stuff
A couple pages were missing $TITLE variables.
Added onClick dropdowns to the list of categories - looks neat, does
absolutely nothing.

By the way the database changed on the last commit I made and I forgot
to mention it. There's now a 'grade' column in the 'courses' table
(which theoretically eventually stores the most recently calculated
overall grade for a course), and a new 'categories' table which contains
id, courseID, name, and weight columns. id is the index and set to
autoincrement.